LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-tokenization-innerheight-innerwidth-expected.txt


FAIL "innerwidth=401" should set width of opened window assert_equals: expected 401 but got 980
FAIL " innerwidth = 401" should set width of opened window assert_equals: expected 401 but got 980
FAIL "innerwidth==401" should set width of opened window assert_equals: expected 401 but got 980
FAIL "\ninnerwidth= 401" should set width of opened window assert_equals: expected 401 but got 980
FAIL ",innerwidth=401,," should set width of opened window assert_equals: expected 401 but got 980
FAIL "INNERWIDTH=401" should set width of opened window assert_equals: expected 401 but got 980
FAIL "innerWidth=401" should set width of opened window assert_equals: expected 401 but got 980
FAIL "innerheight=402" should set height of opened window assert_equals: expected 402 but got 735
FAIL " innerheight = 402" should set height of opened window assert_equals: expected 402 but got 735
FAIL "innerheight==402" should set height of opened window assert_equals: expected 402 but got 735
FAIL "\ninnerheight= 402" should set height of opened window assert_equals: expected 402 but got 735
FAIL ",innerheight=402,," should set height of opened window assert_equals: expected 402 but got 735
FAIL "INNERHEIGHT=402" should set height of opened window assert_equals: expected 402 but got 735
FAIL "innerHeight=402" should set height of opened window assert_equals: expected 402 but got 735
